+++ src/collective/fingerpointing/upgrades/v2/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+from collective.fingerpointing.config import PROJECTNAME
+from collective.fingerpointing.logger import logger
+
+
+def update_configlet(setup_tool):
+    """Update control panel configlet."""
+    profile = 'profile-{0}:default'.format(PROJECTNAME)
+    setup_tool.runImportStepFromProfile(profile, 'controlpanel')
+    logger.info('Control panel configlet updated.')